+++ packages/dara-core/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## NEXT
+
+-  Allow passing a `ClientVariable` to the `DynamicComponent` component
+
 ## 1.21.21
 
 -  Relaxed `fastapi` dependency to `>=0.115.0, <0.121.0`